Create  README.md
📚 Welcome to AWS Cloud Technical Essentials – Week 1
In Week 1, you’ll dive into the fundamentals of cloud computing and explore the value it brings to modern IT systems. You’ll learn to differentiate between on-premises and cloud-based workloads and understand how to create and configure an AWS account. This week also covers a high-level overview of Amazon Web Services (AWS), including AWS Regions and Availability Zones, and introduces the various ways to interact with the AWS platform. Additionally, you’ll learn the importance of AWS Identity and Access Management (IAM) and follow best practices for managing access securely.


🎯 Learning Objectives
✅ Define cloud computing and its value proposition

✅ Differentiate between on-premises vs cloud-based workloads

✅ Describe the core services of Amazon Web Services (AWS)

✅ Understand AWS Regions and Availability Zones

✅ Learn the different ways to interact with AWS

✅ Create your own AWS account

✅ Explore IAM (Identity and Access Management) and its best practices